requirement_tf1.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pyhd8ed1ab_0
astor=0.8.1=pyh9f0ad1d_0
brotli=1.0.9=h7f98852_6
brotli-bin=1.0.9=h7f98852_6
c-ares=1.18.1=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2021.10.8=py37h89c1867_1
cycler=0.11.0=pyhd8ed1ab_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.6=h48d8840_2
expat=2.4.1=h9c3ff4c_0
fontconfig=2.13.1=he4413a7_1000
fonttools=4.28.1=py37h5e8e339_0
freetype=2.10.4=h0708190_1
gast=0.5.3=pyhd8ed1ab_0
gettext=0.19.8.1=h73d1719_1008
glib=2.70.1=h780b84a_0
glib-tools=2.70.1=h780b84a_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.42.0=py37hd5d88b1_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
h5py=3.4.0=nompi_py37hd308b1e_102
hdf5=1.12.1=nompi_h7f166f4_101
icu=58.2=hf484d3e_1000
importlib-metadata=4.8.2=py37h89c1867_0
jbig=2.1=h7f98852_2003
jpeg=9d=h36c2ea0_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.2=py37h2527ec5_1
krb5=1.19.2=h48eae69_3
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=12_linux64_openblas
libbrotlicommon=1.0.9=h7f98852_6
libbrotlidec=1.0.9=h7f98852_6
libbrotlienc=1.0.9=h7f98852_6
libcblas=3.9.0=12_linux64_openblas
libcurl=7.80.0=h494985f_0
libdeflate=1.8=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libglib=2.70.1=h174f98d_0
libgomp=11.2.0=h1d223b6_11
libiconv=1.16=h516909a_0
liblapack=3.9.0=12_linux64_openblas
libnghttp2=1.43.0=ha19adfc_1
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.18=pthreads_h8fe5266_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.19.1=h780b84a_0
libssh2=1.10.0=ha35d2d1_2
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.3.0=h6f004c6_2
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.1=h7f98852_0
libxcb=1.13=h7f98852_1004
libxml2=2.9.9=h13577e0_2
libzlib=1.2.11=h36c2ea0_1013
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.6=pyhd8ed1ab_0
matplotlib=3.5.0=py37h89c1867_0
matplotlib-base=3.5.0=py37h1058ff1_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.2=h58526e2_4
numpy=1.20.3=py37h038b26d_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=hb52868f_1
openssl=3.0.0=h7f98852_2
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.4=py37he8f5f7f_1
patsy=0.5.2=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pillow=8.4.0=py37h0f21c89_0
pip=21.3.1=pyhd8ed1ab_0
protobuf=3.19.1=py37hcd2ae1e_1
pthread-stubs=0.4=h36c2ea0_1001
pydoe=0.3.8=py_1
pyparsing=3.0.6=pyhd8ed1ab_0
pyqt=5.6.0=py37h13b7fb3_1008
python=3.7.12=hf930737_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytz=2021.3=pyhd8ed1ab_0
qt=5.6.3=h8bf5577_3
readline=8.1=h46c0cb4_0
scipy=1.7.2=py37hf2a6cf1_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
setuptools=59.1.1=py37h89c1867_0
sip=4.18.1=py37hf484d3e_1000
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=h9cd32fc_2
statsmodels=0.13.1=py37hb1e94ed_0
tensorboard=1.14.0=py37_0
tensorflow=1.14.0=h4531e10_0
tensorflow-base=1.14.0=py37h4531e10_0
tensorflow-estimator=1.14.0=py37h5ca1d4c_0
termcolor=1.1.0=py_2
tk=8.6.11=h27826a3_1
tornado=6.1=py37h5e8e339_2
typing_extensions=4.0.0=pyha770c72_0
unicodedata2=13.0.0.post2=py37h5e8e339_4
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_1
wrapt=1.13.3=py37h5e8e339_1
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.0=ha95c52a_0
